Better determine response `Result` type
Go through all the methods to find the best response type. Previously,
only the first method determined the response type, but that could be a
method that doesn't return an error. Now, all method return types are
reduced into a single response type. For now this type should be a
common `Result` type, and that assumes that all methods have either the
common `Result` type as the return type, or just the result's wrapped
`Ok` type.