Use an `async` block to create the response future
Instead of using a boxed future for each method call, wrap the whole
request match block inside an `async` block, and await for the
asynchronous methods to finish. This means that now nothing is actually
executed on `Service::call`, and the execution only occurs once the
returned response future is polled.

This also allows using asynchronous code inside the generated code for
each method, which will be necessary for handling different receiver
types.