Add `ReceiverType::service_method_call_prefix`
A method to help with handling the different ways to call a method,
depending on the service receiver type used to determine the service
data and the method receiver type. This generates code that can then be
concatenated with the method name and the method paramaters in order to
call the method.