Return success and error types from `ResultData`
Refactor the getter methods so that instead of generating the code
representing the type, it just returns the information it has. This
means it returns the `syn::Type` representing the `Ok` variant and an
`Option<syn::Type>` representing the `Err` variant if the `ResultData`
represents a `Result` type.

This also moves the code generation from `ResultData` into
`ResponseData`, which also decides how to handle a non-result type.