Create a `Dispatcher` helper type
Keeps a hash map in an `Arc<Mutex<_>>`, so it can be cheaply cloned. The
hash map contains IDs and response endpoints. It can be used either as a
`Sink` of pending response endpoints or as a `Sink` of responses. The
former adds items to the hash map, while the latter removes the endpoint
respective to the response and sends the response to it.